requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.13.4
boto3==1.34.127
botocore==1.34.127
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
curl_cffi==0.11.4
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.2
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
firebase-admin==6.5.0
frozendict==2.4.6
frozenlist==1.4.1
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
greenlet==3.1.1
groq==0.11.0
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.10
langchain-groq==0.2.0
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langsmith==0.1.135
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.51.2
openpyxl==3.1.5
orjson==3.10.5
packaging==24.1
pandas==2.2.3
peewee==3.18.2
pillow==10.3.0
platformdirs==4.3.8
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.5.2
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
pyreadstat==1.2.7
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.30
starlette==0.37.2
tenacity==8.4.1
tesseract==0.1.3
tiktoken==0.7.0
tqdm==4.66.4
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.1
uuid==1.30
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==15.0.1
Werkzeug==3.0.3
yarl==1.9.4
yfinance==0.2.65
